DragEngine/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
ollama
python-docx
pandas
openpyxl
python-pptx
beautifulsoup4
